mv requirements.txt.txt requirements.txt
torch
transformers
torchvision
Pillow
scikit-learn
